utils.py

import time
import asyncio
import json
from typing import Dict, Any, List, Optional, Tuple
import requests
import aiohttp
from solana.rpc.async_api import AsyncClient
from solders.keypair import Keypair
from solders.pubkey import Pubkey
from solders.transaction import VersionedTransaction
from solana.transaction import Transaction
import base58
from loguru import logger

# Set up logging
logger.remove()
logger.add(
    "bot.log",
    rotation="10 MB",
    retention="1 week",
    level="INFO",
    format="{time:YYYY-MM-DD HH:mm:ss.SSS} | {level: <8} | {message}",
)
logger.add(
    lambda msg: print(msg),
    level="INFO",
    format="<green>{time:HH:mm:ss.SSS}</green> | <level>{level: <8}</level> | <cyan>{message}</cyan>",
)

async def create_solana_client(rpc_url: str) -> AsyncClient:
    """Create and return a Solana client."""
    logger.info(f"Connecting to Solana RPC: {rpc_url}")
    return AsyncClient(rpc_url)

def load_keypair(private_key: str) -> Keypair:
    """Load keypair from private key."""
    try:
        secret_key = base58.b58decode(private_key)
        return Keypair.from_bytes(secret_key)
    except Exception as e:
        logger.error(f"Error loading keypair: {str(e)}")
        raise

async def get_token_metadata(client: AsyncClient, token_address: Pubkey) -> Dict[str, Any]:
    """Fetch token metadata from on-chain data."""
    try:
        # This is a simplified approach. In a real implementation,
        # you would query the SPL token metadata program
        token_info = await client.get_token_supply(token_address)
        
        # Basic metadata
        metadata = {
            "address": str(token_address),
            "decimals": 0,  # Default value
            "symbol": "UNKNOWN",
            "name": "Unknown Token",
        }
        
        # Try to get metadata from the Token Metadata program
        try:
            # This is a placeholder - you'd need to implement this
            # based on the actual Token Metadata program
            pass
        except Exception as e:
            logger.warning(f"Could not fetch detailed metadata: {e}")
            
        return metadata
        
    except Exception as e:
        logger.error(f"Error fetching token metadata: {e}")
        return {
            "address": str(token_address),
            "decimals": 0,
            "symbol": "UNKNOWN",
            "name": "Unknown Token",
        }

async def get_jupiter_quote(
    input_token: str,
    output_token: str,
    amount: float,
    slippage: float
) -> Dict[str, Any]:
    """Get a quote from Jupiter Aggregator API."""
    try:
        input_amount = int(amount * (10 ** 6))  # Assuming USDC with 6 decimals
        
        url = f"https://quote-api.jup.ag/v6/quote"
        params = {
            "inputMint": input_token,
            "outputMint": output_token,
            "amount": str(input_amount),
            "slippageBps": int(slippage * 100),  # Convert to basis points
        }
        
        async with aiohttp.ClientSession() as session:
            async with session.get(url, params=params) as response:
                if response.status != 200:
                    error_text = await response.text()
                    logger.error(f"Jupiter quote API error: {error_text}")
                    return None
                
                data = await response.json()
                return data
                
    except Exception as e:
        logger.error(f"Error getting Jupiter quote: {e}")
        return None

async def simulate_jupiter_swap(
    client: AsyncClient,
    keypair: Keypair,
    input_token: str,
    output_token: str,
    amount: float,
    slippage: float
) -> Tuple[bool, Optional[Dict[str, Any]]]:
    """Simulate a swap using Jupiter to check for honeypots or issues."""
    try:
        # Get quote from Jupiter
        quote = await get_jupiter_quote(input_token, output_token, amount, slippage)
        if not quote:
            return False, None
            
        # Get the transaction from Jupiter's swap API
        url = "https://quote-api.jup.ag/v6/swap"
        payload = {
            "quoteResponse": quote,
            "userPublicKey": str(keypair.pubkey()),
            "wrapAndUnwrapSol": True,
        }
        
        async with aiohttp.ClientSession() as session:
            async with session.post(url, json=payload) as response:
                if response.status != 200:
                    error_text = await response.text()
                    logger.error(f"Jupiter swap API error: {error_text}")
                    return False, None
                
                swap_data = await response.json()
                
        # Extract the transaction data for simulation
        transaction_data = swap_data.get("swapTransaction")
        if not transaction_data:
            logger.error("No transaction data in Jupiter response")
            return False, None
        
        # Decode and prepare the transaction for simulation
        tx_bytes = base58.b58decode(transaction_data)
        # For simulation, we don't need to sign
        
        # Simulate the transaction
        sim_result = await client.simulate_transaction(tx_bytes)
        if sim_result.value.err:
            logger.error(f"Simulation failed: {sim_result.value.err}")
            return False, None
            
        # Check for other common issues in the simulation logs
        logs = sim_result.value.logs
        if logs:
            # Check for common error patterns in logs
            error_patterns = ["slippage tolerance exceeded", "insufficient funds", "token balance"]
            for pattern in error_patterns:
                if any(pattern.lower() in log.lower() for log in logs):
                    logger.warning(f"Simulation detected potential issue: {pattern}")
                    return False, None
        
        # If we reached here, the simulation was successful
        return True, quote
        
    except Exception as e:
        logger.error(f"Error simulating swap: {e}")
        return False, None

async def execute_jupiter_swap(
    client: AsyncClient,
    keypair: Keypair,
    quote: Dict[str, Any]
) -> Optional[str]:
    """Execute a swap using Jupiter."""
    try:
        # Get the transaction from Jupiter's swap API
        url = "https://quote-api.jup.ag/v6/swap"
        payload = {
            "quoteResponse": quote,
            "userPublicKey": str(keypair.pubkey()),
            "wrapAndUnwrapSol": True,
        }
        
        async with aiohttp.ClientSession() as session:
            async with session.post(url, json=payload) as response:
                if response.status != 200:
                    error_text = await response.text()
                    logger.error(f"Jupiter swap API error: {error_text}")
                    return None
                
                swap_data = await response.json()
                
        # Extract and sign the transaction
        transaction_data = swap_data.get("swapTransaction")
        if not transaction_data:
            logger.error("No transaction data in Jupiter response")
            return None
            
        tx_bytes = base58.b58decode(transaction_data)
        # In a real implementation, you would properly deserialize, sign and serialize
        # This is a placeholder for the actual implementation
        
        # Sign and send the transaction
        # Note: This is a simplified placeholder. You'll need to implement
        # proper transaction signing based on how Jupiter formats their transactions
        # (VersionedTransaction or legacy Transaction)
        tx_sig = await client.send_transaction(tx_bytes, keypair)
        
        # Return the transaction signature
        return tx_sig.value
        
    except Exception as e:
        logger.error(f"Error executing swap: {e}")
        return None

async def check_token_price(
    input_token: str,
    output_token: str,
    amount: float
) -> float:
    """Check current token price using Jupiter quote API."""
    try:
        # For price check we use a fixed amount to get current rate
        input_amount = int(amount * (10 ** 6))  # Assuming USDC with 6 decimals
        
        url = f"https://quote-api.jup.ag/v6/quote"
        params = {
            "inputMint": input_token,
            "outputMint": output_token,
            "amount": str(input_amount),
            "slippageBps": 50,  # Small slippage for price check
        }
        
        async with aiohttp.ClientSession() as session:
            async with session.get(url, params=params) as response:
                if response.status != 200:
                    logger.error("Failed to get price quote")
                    return 0
                
                data = await response.json()
                
                # Calculate the price
                output_amount = int(data.get("outAmount", 0))
                if output_amount == 0:
                    return 0
                    
                output_decimals = int(data.get("outputDecimals", 0))
                price = output_amount / (10 ** output_decimals) / amount
                
                return price
                
    except Exception as e:
        logger.error(f"Error checking token price: {e}")
        return 0

def is_honeypot(quote: Dict[str, Any]) -> bool:
    """Check if a token might be a honeypot based on Jupiter quote."""
    # Check for extremely high price impact
    price_impact = float(quote.get("priceImpactPct", 0)) * 100
    if price_impact > 15:
        logger.warning(f"High price impact detected: {price_impact}%")
        return True
        
    # Check for other potential red flags
    routes = quote.get("routesInfos", [])
    if not routes:
        return True
    
    # Additional checks could be added here
    
    return False
