requirements.txt

solana==0.31.1
solders>=0.22.0
python-dotenv==1.0.1
aiohttp==3.9.3
requests==2.31.0
loguru==0.7.2
python-telegram-bot==20.7
pydantic==2.6.3
websockets==12.0
